requirements.txt
beautifulsoup4==4.12.3
google-generativeai==0.8.3
jupyter==1.1.1
jupyterlab==4.2.6
load-dotenv==0.1.0
lxml==5.3.0
matplotlib==3.9.2
openai==1.55.0
pandas==2.2.3
requests==2.32.3
seaborn==0.13.2
statsmodels==0.14.4
streamlit==1.40.1
pydantic
numpy
wheel
fastparquet
pyarrow
sentence_transformers
faiss-cpu
joblib